Split SyntaxNodeExtensions file into Documentation extension
